Change logic around `ctypes` addresses
The `ctypes` addresses were displaying strange behaviour as it attempted to
perform `strlen` on the values. This replaces them with `c_size_t` to avoid
having any memory probing. This also changed the `Value` handle to use raw
pointer to `PyObject*` rather than the `ctypes.py_object`, which do wrapping.